cleanup claims (#20)

+++ data_generator/util/assessment_stats.py
@@ -225,9 +225,9 @@ def random_claims(score: int, claim_weights: [float], claim_min: int, claim_max:
         remaining_weight -= claim_weight
 
         min_ = min(claim_max,
-                   max(claim_min, math.floor((remaining_score - remaining_weight * claim_max) / claim_weight)))
+                   max(claim_min, int(math.floor((remaining_score - remaining_weight * claim_max) / claim_weight))))
         max_ = max(claim_min,
-                   min(claim_max, math.ceil((remaining_score - remaining_weight * claim_min) / claim_weight)))
+                   min(claim_max, int(math.ceil((remaining_score - remaining_weight * claim_min) / claim_weight))))
 
         assert min_ <= max_, "%s %s" % (min_, max_)
